Add all but cup.gdb data files and create zip script
Created a powershell script to compress and expand data files as needed. The compressed files are added to the repository in this commit. Various smaller, uncompressed data and example files are also added in this commit. The cup.gdb directory is not included in this commit due to the compressed file still exceeding 100MB. More work needs to be done for this. The offending file is a00000009.gdbtable ~ 491MB uncompressed.