Update script that processes the withdrawl point input file
In main script, redirected the preprocess directory to the results directory instead of the now defunct preprocess directory. The input file processing script uses this redirection, as well as the linking of specific files that are pointed to in the preprocess directory. Also, the script now receives and utilizes the conversion factor mgd2cfd from the definitions file.