Replace remaining inflated gis files/folders with zipped versions
Replaced the projection folder and the two mxd files; dh.mxd and temp_updatedLinks.mxd with two zip files: mxd.zip and projections.zip. This is intended to reduce the size of data in the repository, speed up repository download, and to allow an easier avenue to copy necessary files to a results directory during tool runtime. Later changes will reduce the size of copied files to the results folder