Reorganization of sim_cup_main.py
Made significant changes to reorganize where files are read from and where they are written to.

The changes implement the need to create a new results directory for each run, and also corrects issues with some files not being available usable from a central definitions_files_directory.

Now each run is completely separate from the previous, including the gis results. This is a first step in this kind of reorganization, where everything used during runtime is copied to a newly created (at runtime) results directory. Later, files that can be kept in a centralized directory will not be copied.