Correct PATHs in sim_cup_main.py
Made corrections to how the current_working_directory PATH is extracted using join with splat (*), as well as extracting the parent directory using  pardir and abspath. The fix is not fully OS independent, though, because there is still a split based on ('\\'). This will need to be addressed later.

The basename filename is also fixed using ntpath.basename to extract the filename from any PATH, even when the file is in the current directory and no additional PATH is given. Now the input file can reside in its own directory if the user wishes.