Tmp change to search for the parrent tool directory, again
Once again had to change how the current working directory is parsed in order to get to the parent directory. The currrent change uses os.path.split instead of os.path.abspath with os.pardir to define the results main directory.

In console testing the abspath method worked well. But during runtime the results main directory actually added both the current working and results directories PATHs.  For now this has been changed to simply remove the last item in the list of PATH components. This works for now, but will not be robust in the long run. An issue should be raised to address this.