Reverting to previous CWD retrieval
Had to revert to a previous version of retrieving the currrent working directory. Testing of other (recommended) methods succeeded, but during runtime would fail. The runtime failure may be associated with a "\n" in the PATH due to the name nfseg. The Windows backslash turns this into a newline character. PATHs used during testing didn't have this. This current implementation reverted back to replacing the Windows-style \ with unix-style / .

Another error addressed was the deletion of the results directory when it was not empty. Replaced os.rmdir with shutil.rmtree