Organized example input files and updated .gitignore
Moved the remaining example input files to the user_input_files folder. Also modified the second sjrwmd example: the X,Y coordinates the same as the first example (verified coordinates) while all other information, including the flow rate, is from the srwmd example (srwmd currently does not work).

In addition, the .gitignore file has been updated for which files are not included in the repository.