Added starting heads file back into model dir
Inadvertently removed the starting heads file. It has been put back. All the files in the model directory have now been checked against those listed in the .nam file. There remains extra files that are not related to the files needed for runtime (modflow or pest), but these files are small and related to what comes from the published model version. It is possible these get removed later, but remain for now.