Removed unused
Removed unused shape file components in wellpkg_update and updated the zip file. The zip file was reduced from ~82MB to ~3MB.

Added logfile print support to the update_wellpkg_nfseg.

Added statement to cleanup the Feature Layer memory. This hopefully addresses the error that arises when running multiple input files in the while loop, that a FeatureLayer already exists.