Remove unnecessary (and incorrect) if statement that prevented cfd-to-cfs conversion
Removed an if statement conditional on the number of main function arguments. As it was implemented, if the condition was met (number of arguments matched specified number) the conversion of cfd-to-cfs was applied, otherwise not. But the specified number of main function arguments was incorrect, meaning the condition was never met. Turns out the if statement was not needed at all anymore, so just removed the if statement and potential to get it wrong again. The conversion happens every time now.