Remove dh directory accidentally added to repository
A single file was added to the repository in the postproc/dh directory, and this caused the directory to show up in the repository before it was ready. Removing this file will remove the directory. A later change will add the directory again, but reformated.

The removed file is part of what is extracted from the dh.zip, so is still part of the repository.