Switched to using new heads processing routine
Modified the program to now use the newly implemented heads processing routines.

The previous commit implemented the placement and organization for the new routine. This commit made the full switch to use the new routine to extract and process heads for all model layers, and replace the old method that used PEST routines for only model layers 1,3. The implementation has been tested, and for model layers 1,3 the change in heads are approximately same as previous results (not exact because of less rounding). Code lines associated with the older method are commented out in the MAIN program for now, and will require cleaning, and can be fully removed once more testing by more people has been completed.

Also changed, the postprocessing/dh directory no longer uses a zipped file. The dh.zip file from input_and_definition_files/postprocess/dh.zip file was replaced with the dh directory in the same place. Rather than unzipping all the contents into the results/postproc/dh directory, the dh directory is made and only the needed files are copied to it. The reference dh directory in input_and_definitions_files now includes the input_control_file for the new routines, as well as retaining (for now) the PEST related input files for the old method of processing heads. All output files

The new routine uses a Fortran executable to extract heads and for all model layers process the change in heads (dh) between stress periods 2 minus 1, and calculate the area-averaged lake heads and dh. The extracted heads and dh are output to file for all layers.

Furture changes need to process the lake heads output.