Attempt at fixing intermittent issue of replacing the results directory
Sometimes when replacing an existing results directory, the os.mkdir command in MAIN fails because the directory did not finish deleting yet (throws error that directory exists). The mkdir command is now in a try-except statement to provide a second chance at removing / making the directory before throwing an error. This will help to slow the system down just long enough (hopefully) to allow the directory to fully delete.