Revamped cup well gdb processing
Cleaned up the process for locating cup well coordinates within the NFSEG grid.

Previously, the cup well locations were saved to wellpkg_update.gdb, and then those contents were copied to the cup.gdb.

Now, the cup well locations are saved directly to the cup.gdb without using the wellpkg_update.gdb as an intermediate step. The wellpkg_update.gdb served no other purpose after this change, and was removed.

Modified the inputs to the script that processes and locates the cup wells on the grid to now directly accept the grid and map projection filenames and PATHs in order to reduce duplication (filenames and PATHs for map projections were defined in both this processing script and MAIN).

Renamed the cup well processing/locating script to show evolutionary history.