Organized and completed the collection of heads extraction input files
Added any missing input files to the lakes portion of the heads extraction utility. Also made sure all files within this utility matched those in the archived utility directory (was in the repository directory but ignored). The archived directory was renamed and moved to a different repository of general NFSEG processing utilities.

The gitignore file was updated to reflect the changes.