Upload dQ directory files
Uploaded a cleaned and uncompressed version of the dQ directory. This will replace the dQ.zip file currently being used, which be removed in a subsequent commit along with a change to the MAIN script to use the uncompressed version.

The archived files will stored in a separate location outside of the repository