Remove archived dH files from the active directory
Removed files that were input files for a previous version of the Reg Tool. These files are no longer used by the Tool, but are archived to keep a record of past input files that were also used for NFSEG v1.1 calibration in some capacity. The archived folder will be move to a separate location outside of the repository